introduce parallel runs for commands
